Playground: Implement Hot Module Replacement.
