+++ playground/src/index.js
@@ -1,30 +1,17 @@
 /**
- * WordPress dependencies
+ * External dependencies
  */
-import '@wordpress/editor'; // This shouldn't be necessary
-
-import { render, useState, Fragment } from '@wordpress/element';
-import {
-	BlockEditorKeyboardShortcuts,
-	BlockEditorProvider,
-	BlockList,
-	BlockInspector,
-	WritingFlow,
-	ObserveTyping,
-} from '@wordpress/block-editor';
-import {
-	Button,
-	Popover,
-	SlotFillProvider,
-	DropZoneProvider,
-} from '@wordpress/components';
-import { registerCoreBlocks } from '@wordpress/block-library';
-import '@wordpress/format-library';
+import 'react-hot-loader';
 
 /**
- * Internal dependencies
+ * WordPress dependencies
  */
-import './style.scss';
+// This shouldn't be necessary
+import '@wordpress/editor';
+import '@wordpress/format-library';
+
+import { registerCoreBlocks } from '@wordpress/block-library';
+import { render } from '@wordpress/element';
 
 /* eslint-disable no-restricted-syntax */
 import '@wordpress/components/build-style/style.css';
@@ -35,44 +22,16 @@ import '@wordpress/block-library/build-style/theme.css';
 import '@wordpress/format-library/build-style/style.css';
 /* eslint-enable no-restricted-syntax */
 
-function App() {
-	const [ blocks, updateBlocks ] = useState( [] );
+/**
+ * Internal dependencies
+ */
+import App from './app';
 
-	return (
-		<Fragment>
-			<div className="playground__header">
-				<h1 className="playground__logo">Gutenberg Playground</h1>
-				<Button isLarge href="design-system/components" target="_blank">
-					Design System Components
-				</Button>
-			</div>
-			<div className="playground__body">
-				<SlotFillProvider>
-					<DropZoneProvider>
-						<BlockEditorProvider
-							value={ blocks }
-							onInput={ updateBlocks }
-							onChange={ updateBlocks }
-						>
-							<div className="playground__sidebar">
-								<BlockInspector />
-							</div>
-							<div className="editor-styles-wrapper">
-								<BlockEditorKeyboardShortcuts />
-								<WritingFlow>
-									<ObserveTyping>
-										<BlockList />
-									</ObserveTyping>
-								</WritingFlow>
-							</div>
-							<Popover.Slot />
-						</BlockEditorProvider>
-					</DropZoneProvider>
-				</SlotFillProvider>
-			</div>
-		</Fragment>
-	);
-}
+import './style.scss';
 
 registerCoreBlocks();
 render( <App />, document.querySelector( '#app' ) );
+
+if ( module.hot ) {
+	module.hot.accept();
+}
